edgerules: move shared definitions to a edgerules.go file
Move definitions that are used by the pullzone and edgerule endpoints to a
edgerules.go file.
This makes those easier to find then in pullzone_get.go.
